requirements-airflow.txt
kafka-python
pyspark
snowflake-connector-python
requests
python-dotenv